updated swaggers, added AdditionalImage

+++ favorites/views.py
@@ -15,37 +15,31 @@ class FavoriteProductsToggleView(APIView):
     @swagger_auto_schema(
         operation_description="Toggle favorite product.",
         responses={
-            200: "Product removed from favorites.",
-            201: "Product added to favorites.",
+            201: '{"message": "Product added to favorites"}',
+            200: '{"message": "Product removed from favorites"}',
             403: "Forbidden. The user does not have the required permissions.",
+            404: "Product or user not found."
         }
     )
-
     def post(self, request, product_id):
         try:
             product = Product.objects.get(pk=product_id)
         except Product.DoesNotExist:
             return Response(
                 {"message": "Product not found"}, status=status.HTTP_404_NOT_FOUND)
-        try:
-            user = CustomUser.objects.get(pk=request.user.id)
-            print(user)
-            print(request.user.id)
-        except CustomUser.DoesNotExist:
-            return Response(
-                {"message": "User not found"}, status=status.HTTP_404_NOT_FOUND)
 
-        try:
-            favorite_product = FavoriteProducts.objects.get(
-                product=product, user=user)
+        user = request.user
+        favorite_product, created = FavoriteProducts.objects.get_or_create(product=product, user=user)
+        
+        if created:
+            return Response(
+                {"message": "Product added to favorites"}, status=status.HTTP_201_CREATED)
+        else:
             favorite_product.delete()
             return Response(
                 {"message": "Product removed from favorites"}, status=status.HTTP_200_OK)
-        except FavoriteProducts.DoesNotExist:
-            favorite_product = FavoriteProducts(product=product, user=user)
-            favorite_product.save()
-            return Response(
-                {"message": "Product added to favorites"}, status=status.HTTP_201_CREATED)
+
+    permission_classes = [IsAuthenticated]
 
 
 class FavoriteProductsListView(APIView):
@@ -53,8 +47,9 @@ class FavoriteProductsListView(APIView):
     @swagger_auto_schema(
         operation_description="List all favorite products.",
         responses={
-            200: "List of favorite products.",
+            200: FavoriteProductsSerializer,
             403: "Forbidden. The user does not have the required permissions.",
+            404: "User not found.",
         }
     )
     def get(self, request):
@@ -63,27 +58,18 @@ class FavoriteProductsListView(APIView):
         except CustomUser.DoesNotExist:
             return Response(
                 {"message": "User not found"}, status=status.HTTP_404_NOT_FOUND)
-        favorite_products = FavoriteProducts.objects.filter(user=user)
-        serializer = FavoriteProductsSerializer(favorite_products, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
 
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request):
-        try:
-            user = CustomUser.objects.get(pk=request.user.id)
-        except CustomUser.DoesNotExist:
-            return Response(
-                {"message": "User not found"}, status=status.HTTP_404_NOT_FOUND)
         favorite_products = FavoriteProducts.objects.filter(user=user)
         serializer = FavoriteProductsSerializer(favorite_products, many=True)
         favorite_products_data = serializer.data
 
-        favorite_products = {}
+        favorite_products_dict = []
         for favorite_product_data in favorite_products_data:
             product_id = favorite_product_data['product']
             product = Product.objects.get(pk=product_id)
             product_serializer = ProductCreateSerializer(product)
-            favorite_products[product_id] = product_serializer.data
+            product_data = product_serializer.data
+            favorite_products_dict.append(product_data)
+        return Response(favorite_products_dict, status=status.HTTP_200_OK)
 
-        return Response(favorite_products, status=status.HTTP_200_OK)
+    permission_classes = [IsAuthenticated]
